requirements.txt
supabase==2.19.0
postgrest==2.19.0
realtime==2.19.0
storage3==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
httpx==0.27.0
python-dotenv>=1.0.1
requests>=2.31.0
streamlit==1.37.1
pandas==2.2.2
openpyxl>=3.1.0
# Pin to satisfy Streamlit dependency constraints
packaging==24.1
protobuf==4.25.3
